greenblock-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
requests==2.31.0
python-dotenv==1.0.1